backend/run.py
"""LUMIS Audit Backend - Entry Point."""
import os
from dotenv import load_dotenv

load_dotenv()

from app import create_app

app = create_app()

if __name__ == "__main__":
    host = os.getenv("FLASK_HOST", "0.0.0.0")
    port = int(os.getenv("FLASK_PORT", "5002"))
    debug = os.getenv("FLASK_DEBUG", "false").lower() == "true"
    print(f"\n[LUMIS] Audit backend starting on http://{host}:{port}\n")
    app.run(host=host, port=port, debug=debug, threaded=True)
