backend/requirements.txt
flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.1
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.54.0
requests==2.32.3
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
faker==30.8.0
reportlab==4.2.5
supabase==2.9.1
google-cloud-pubsub==2.26.1
google-cloud-aiplatform==1.71.1
google-maps-platform-datasets==0.3.8
